requirements.txt
aiofiles==23.2.1
aiosmtplib==3.0.1
alembic==1.13.0
asgiref==3.7.2
asyncmy==0.2.9
bcrypt==4.0.1
casbin==1.33.0
casbin-async-sqlalchemy-adapter==1.3.0
celery==5.3.6
cryptography==41.0.7
email-validator==2.0.0
fast-captcha==0.2.1
fastapi==0.105.0
fastapi-limiter==0.1.5
fastapi-pagination==0.12.13
gunicorn==21.2.0
httpx==0.25.2
itsdangerous==2.1.2
loguru==0.7.2
passlib==1.7.4
path==15.1.2
phonenumbers==8.13.27
pre-commit==3.2.2
psutil==5.9.6
pydantic==2.5.2
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pytest==7.2.2
pytest-pretty==1.2.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
redis[hiredis]==4.5.5
ruff==0.1.8
SQLAlchemy==2.0.23
starlette==0.27.0
supervisor==4.2.5
user-agents==2.2.0
uvicorn[standard]==0.24.0
wait-for-it==2.2.2
XdbSearchIP==1.0.2

typing_extensions~=4.9.0